% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HappyCampR.r
\docType{package}
\name{HappyCampR}
\alias{HappyCampR}
\title{Shiny App to Find Campgrounds}
\description{
Shiny App to Find Campgrounds
}
\note{
This package is maintained as a hobby.
}
\section{Legal Mumbo Jumbo}{


HappyCampR is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU Lesser General Public License for more details.
}

\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\keyword{package}
