% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HappyCampR.r
\docType{data}
\name{MoreCamp}
\alias{MoreCamp}
\title{Campground Data}
\format{A \code{data.frame} object with XXX rows and XX columns. The
data are from

The columns are defined as follows:
\describe{
 \item{\code{amt}}{The numeric amount of the ingredient.}
}}
\source{
place, \url{campsite.info}
}
\usage{
MoreCamp
}
\description{
...
}
\examples{
\dontrun{
data(MoreCamp)

}
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\keyword{datasets}
