% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HappyCampR.r
\name{HappyCampR-NEWS}
\alias{HappyCampR-NEWS}
\title{News for package 'HappyCampR':}
\description{
News for package \sQuote{HappyCampR}

\newcommand{\CRANpkg}{\href{https://cran.r-project.org/package=#1}{\pkg{#1}}}
\newcommand{\HappyCampR}{\CRANpkg{HappyCampR}}
}
\section{\HappyCampR{} Initial Version 0.1.0 (2018-06-12) }{

\itemize{
\item first CRAN release.
}
}

