% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campr_app.r
\name{campr_app}
\alias{campr_app}
\title{campr_app .}
\usage{
campr_app()
}
\value{
Runs the shiny app.
}
\description{
One sentence or so that tells you some more.
}
\details{
Really detailed. \eqn{\zeta}{zeta}.

A list:
\itemize{
\item I use \eqn{n}{n} to stand for blah.
\item and so forth....
}
}
\examples{
\dontrun{
campr_app()
}
}
\references{
Johnson, N. L., and Welch, B. L. "Applications of the non-central t-distribution."
Biometrika 31, no. 3-4 (1940): 362-389. \url{http://dx.doi.org/10.1093/biomet/31.3-4.362}
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}

Steven E. Pav \email{steven@gilgamath.com}
}
\keyword{shiny}
