% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campr_app.r
\name{campr_app}
\alias{campr_app}
\title{campr_app .}
\usage{
campr_app()
}
\value{
Runs the shiny app.
}
\description{
One sentence or so that tells you some more.
}
\details{
Really detailed. \eqn{\zeta}{zeta}.

A list:
\itemize{
\item I use \eqn{n}{n} to stand for blah.
\item and so forth....
}
}
\examples{
\dontrun{
campr_app()
}
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}

Steven E. Pav \email{steven@gilgamath.com}
}
\keyword{shiny}
