% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campr_app.r
\name{campr_app}
\alias{campr_app}
\title{campr_app .}
\usage{
campr_app(init_lat = 37.7749, init_lon = -122.4194,
  init_search = "San Francisco")
}
\arguments{
\item{init_lat}{an initial latitude to seed the search.}

\item{init_lon}{an initial longitude to seed the search. The
default values are for San Francisco.}

\item{init_search}{an initial string to put into the search text.}
}
\value{
a shiny app.
}
\description{
A shiny app to locate campgrounds.
}
\examples{
\dontrun{
campr_app()
}
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}

Steven E. Pav \email{shabbychef@gmail.com}
}
\keyword{shiny}
